import zmq
context = zmq.Context()

print("Connecting to hello world server...")
socket = context.socket(zmq.REQ)
socket.connect("tcp://localhost:5555")

for request in range(10):
    print(f"Sending request {request} ...")
    socket.send_string("Hello")

    #  Get the reply.
    message = socket.recv()
    print(f"Received reply {request} [ {message} ]")

+++ lab0/problem3.c
@@ -3,8 +3,8 @@ int main () {
     int a = 125, b = 12345; 
     long ax = 1234567890; 
     short s = 4043; 
-    float x = 2.13459; 
-    double dx = 1.1415927; 
+    float x = (int)2.13459; 
+    double dx = (int)1.1415927; 
     char c = 'W'; 
     unsigned long ux = 2541567890;
